+++ VectorMath/Globals.cs
@@ -4,9 +4,9 @@ namespace VectorMath
 {
     public struct Constants
     {
-        public const decimal Eps = 1e-24m;
+        public const double Eps = 1e-24m;
     }
 
 
-    public delegate decimal AppliableFunction(decimal x);
+    public delegate double AppliableFunction(double x);
 }
